% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyInfinite.R
\name{anyInfinite}
\alias{anyInfinite}
\title{Check if an object contains infinite values}
\usage{
anyInfinite(x)
}
\arguments{
\item{x}{[\code{ANY}]\cr
Object to check.}
}
\value{
[\code{logical(1)}] Returns \code{TRUE} if any element is \code{-Inf} or \code{Inf}.
}
\description{
Supported are atomic types (see \code{\link[base]{is.atomic}}), lists and data frames.
}
\examples{
anyInfinite(1:10)
anyInfinite(c(1:10, Inf))
iris[3, 3] = Inf
anyInfinite(iris)
}

