% Generated by roxygen2 (4.0.1): do not edit by hand
\name{asInteger}
\alias{asInteger}
\title{Convert an argument to an integer}
\usage{
asInteger(x, ..., tol = .Machine$double.eps^0.5, .var.name)
}
\arguments{
\item{x}{[ANY]\cr
Object to convert.}

\item{...}{[ANY]\cr
Additional arguments passed to \code{\link{checkInteger}}.}

\item{tol}{[\code{double(1)}]\cr
Numerical tolerance used to check if a double or complex can be converted.
Default is \code{sqrt(.Machine$double.eps)}.}
}
\value{
Converted \code{x}.
}
\description{
Convertible are (a) atomic vectors with all elements \code{NA}
and (b) double vectors with all elements being within \code{tol}
range of an integer.
}

