% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_access}
\alias{check_access}
\title{Check file system access rights}
\usage{
check_access(x, access)
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{access}{[\code{character}]\cr
Single string with characters \sQuote{r}, \sQuote{w} and \sQuote{x} to
force a check for read, write or execute access rights.}
}
\value{
Returns \code{TRUE} on success and an error message otherwise.
 Intended to be used with \code{\link{test}}, \code{\link{assert}} or \code{\link{pass}}.
}
\description{
Check file system access rights
}
\examples{
# Is R's home directory readable?
 test(R.home(), "access", "r")

 # Is R's home directory writeable?
 test(R.home(), "access", "w")
}
\seealso{
Other filesystem: \code{\link{check_directory}};
  \code{\link{check_file}}
}

