% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_choice}
\alias{check_choice}
\title{Check if an object is an element of a given set}
\usage{
check_choice(x, choices)
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{choices}{[\code{atomic}]\cr
Set of possible values.}
}
\value{
Returns \code{TRUE} on success and an error message otherwise.
 Intended to be used with \code{\link{test}}, \code{\link{assert}} or \code{\link{pass}}.
}
\description{
Check if an object is an element of a given set
}
\examples{
test("x", "choice", letters)
}
\seealso{
Other set: \code{\link{check_subset}}
}

