% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_constant}
\alias{check_constant}
\title{Check if all elements of a vector are equal / constant}
\usage{
check_constant(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{tol}{[\code{double(1)}]\cr
Numerical tolerance used if \code{x} is of type \code{double} or \code{complex}.
Default is \code{sqrt(.Machine$double.eps)}.}
}
\value{
Returns \code{TRUE} on success and an error message otherwise.
 Intended to be used with \code{\link{test}}, \code{\link{assert}} or \code{\link{pass}}.
}
\description{
Note that zero-length input is treated as constant.
}
\examples{
test(c(0, 1-0.9-0.1), "constant")
 # should be FALSE due to rounding errors
 test(c(0, 1-0.9-0.1), "constant", tol = 0)
}
\seealso{
Other constant: \code{\link{check_inconstant}}
}

