% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_directory}
\alias{check_directory}
\title{Check for existence and access rights of directories}
\usage{
check_directory(x, access = "")
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{access}{[\code{character}]\cr
Single string with characters \sQuote{r}, \sQuote{w} and \sQuote{x} to
force a check for read, write or execute access rights.}
}
\value{
Returns \code{TRUE} on success and an error message otherwise.
 Intended to be used with \code{\link{test}}, \code{\link{assert}} or \code{\link{pass}}.
}
\description{
Check for existence and access rights of directories
}
\examples{
# Is R's home directory readable?
 test(R.home(), "access", "r")

 # Is R's home directory writable?
 test(R.home(), "access", "w")
}
\seealso{
Other filesystem: \code{\link{check_access}};
  \code{\link{check_file}}
}

