% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_integer}
\alias{check_integer}
\title{Check if an argument is vector of type integer}
\usage{
check_integer(x, lower, upper, ...)
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{lower}{[\code{numeric(1)}]\cr
Lower value all elements of \code{x} must be greater than or equal.}

\item{upper}{[\code{numeric(1)}]\cr
Upper value all elements of \code{x} must be lower than or equal.}

\item{...}{[ANY]\cr
Additional parameters used in a call of \code{\link{check_vector}}.}
}
\value{
Returns \code{TRUE} on success and an error message otherwise.
 Intended to be used with \code{\link{test}}, \code{\link{assert}} or \code{\link{pass}}.
}
\description{
Check if an argument is vector of type integer
}
\details{
This function does not distinguish between
\code{NA}, \code{NA_integer_}, \code{NA_real_}, \code{NA_complex_}
and \code{NA_character_}.
}
\examples{
test(1L, "integer")
 test(1., "integer")
 test(1:2, "integer", lower = 1, upper = 2, any.missing = FALSE)
}
\seealso{
Other basetypes: \code{\link{check_character}};
  \code{\link{check_complex}};
  \code{\link{check_data.frame}};
  \code{\link{check_environment}};
  \code{\link{check_factor}}; \code{\link{check_function}};
  \code{\link{check_integerish}}; \code{\link{check_list}};
  \code{\link{check_logical}}; \code{\link{check_matrix}};
  \code{\link{check_numeric}}; \code{\link{check_vector}}
}

