% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogenerated.r, R/checkAtomic.r
\name{assertAtomic}
\alias{assertAtomic}
\alias{assert_atomic}
\alias{checkAtomic}
\alias{expect_atomic}
\alias{testAtomic}
\alias{test_atomic}
\title{Check that an argument is an atomic vector}
\usage{
assertAtomic(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  add = NULL, .var.name)

assert_atomic(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  add = NULL, .var.name)

testAtomic(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL)

test_atomic(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL)

expect_atomic(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  info = NULL, label = NULL)

checkAtomic(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL)
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{any.missing}{[\code{logical(1)}]\cr
Are vectors with missing values allowed? Default is \code{TRUE}.}

\item{all.missing}{[\code{logical(1)}]\cr
Are vectors with only missing values allowed? Default is \code{TRUE}.}

\item{len}{[\code{integer(1)}]\cr
Exact expected length of \code{x}.}

\item{min.len}{[\code{integer(1)}]\cr
Minimal length of \code{x}.}

\item{max.len}{[\code{integer(1)}]\cr
Maximal length of \code{x}.}

\item{unique}{[\code{logical(1)}]\cr
Must all values be unique? Default is \code{FALSE}.}

\item{names}{[\code{character(1)}]\cr
Check for names. See \code{\link{checkNamed}} for possible values.
Default is \dQuote{any} which performs no check at all.
Note that you can use \code{\link{checkSubset}} to check for a specific set of names.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertions. See \code{\link{AssertCollection}}.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}

\item{info}{[character(1)]\cr
Extra information to be included in the message for the testthat reporter.
See \code{\link[testthat]{expect_that}}.}

\item{label}{[character(1)]\cr
Same as \code{.var.name}, but passed down to \code{\link[testthat]{expect_that}}.}
}
\value{
Depending on the function prefix:
 If the check is successful, all functions return \code{TRUE}. If the check
 is not successful, \code{assertAtmoic}/\code{assert_atmoic}
 throws an error message, \code{testAtmoic}/\code{test_atmoic}
 returns \code{FALSE},
 and \code{checkAtmoic} returns a string with the error message.
 The function \code{expect_atmoic} will return an
 \code{\link[testthat]{expectation}}.
}
\description{
For the definition of \dQuote{atomic}, see \code{\link[base]{is.atomic}}.
}
\examples{
testAtomic(letters, min.len = 1L, any.missing = FALSE)
}
\seealso{
Other atomicvector: \code{\link{assertAtomicVector}},
  \code{\link{assertVector}}

Other basetypes: \code{\link{assertArray}},
  \code{\link{assertAtomicVector}},
  \code{\link{assertCharacter}},
  \code{\link{assertComplex}},
  \code{\link{assertDataFrame}},
  \code{\link{assertEnvironment}},
  \code{\link{assertFactor}}, \code{\link{assertFunction}},
  \code{\link{assertIntegerish}},
  \code{\link{assertInteger}}, \code{\link{assertList}},
  \code{\link{assertLogical}}, \code{\link{assertMatrix}},
  \code{\link{assertNumeric}}, \code{\link{assertVector}}

Other basetypes: \code{\link{assertArray}},
  \code{\link{assertAtomicVector}},
  \code{\link{assertCharacter}},
  \code{\link{assertComplex}},
  \code{\link{assertDataFrame}},
  \code{\link{assertEnvironment}},
  \code{\link{assertFactor}}, \code{\link{assertFunction}},
  \code{\link{assertIntegerish}},
  \code{\link{assertInteger}}, \code{\link{assertList}},
  \code{\link{assertLogical}}, \code{\link{assertMatrix}},
  \code{\link{assertNumeric}}, \code{\link{assertVector}}
}

