% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkCharacter}
\alias{assertCharacter}
\alias{checkCharacter}
\alias{testCharacter}
\title{Check if an argument is a vector of type character}
\usage{
checkCharacter(x, pattern = NULL, ignore.case = FALSE, perl = FALSE,
  fixed = FALSE, min.chars = 0L, ...)

assertCharacter(x, pattern = NULL, ignore.case = FALSE, perl = FALSE,
  fixed = FALSE, min.chars = 0L, ..., .var.name)

testCharacter(x, pattern = NULL, ignore.case = FALSE, perl = FALSE,
  fixed = FALSE, min.chars = 0L, ...)
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}

\item{pattern}{[\code{character(1L)}]\cr
Regular expression as used to use in \code{\link[base]{grepl}}.
All elements of \code{x} must comply to this pattern.
Defaults to \code{NA}.}

\item{ignore.case}{[\code{logical(1)}]\cr
See \code{\link[base]{grepl}}. Default is \code{FALSE}.}

\item{perl}{[\code{logical(1)}]\cr
See \code{\link[base]{grepl}}. Default is \code{FALSE}.}

\item{fixed}{[\code{logical(1)}]\cr
See \code{\link[base]{grepl}}. Default is \code{FALSE}.}

\item{min.chars}{[\code{integer(1)}]\cr
Minimum number of characters in each element of \code{x}.}

\item{...}{[ANY]\cr
Additional parameters used in a call of \code{\link{checkVector}}.}
}
\value{
Depending on the function prefix:
 If the check is successfull, all functions return \code{TRUE}.
 If the check is not successfull, \code{assertCharacter} throws an error message,
 \code{testCharacter} returns \code{FALSE} and \code{checkCharacter} returns
 a string with the error message.
}
\description{
Check if an argument is a vector of type character
}
\details{
This function does not distinguish between
\code{NA}, \code{NA_integer_}, \code{NA_real_}, \code{NA_complex_}
and \code{NA_character_}.
}
\examples{
test(letters, "character", min.len = 1, any.missing = FALSE)
 test(letters, "character", min.chars = 2)
 test("example", "character", pattern = "ple")
}
\seealso{
Other basetypes: \code{\link{assertComplex}},
  \code{\link{checkComplex}}, \code{\link{testComplex}};
  \code{\link{assertDataFrame}},
  \code{\link{checkDataFrame}},
  \code{\link{testDataFrame}};
  \code{\link{assertEnvironment}},
  \code{\link{checkEnvironment}},
  \code{\link{testEnvironment}};
  \code{\link{assertFactor}}, \code{\link{checkFactor}},
  \code{\link{testFactor}}; \code{\link{assertFunction}},
  \code{\link{checkFunction}}, \code{\link{testFunction}};
  \code{\link{assertIntegerish}},
  \code{\link{checkIntegerish}},
  \code{\link{testIntegerish}};
  \code{\link{assertInteger}}, \code{\link{checkInteger}},
  \code{\link{testInteger}}; \code{\link{assertList}},
  \code{\link{checkList}}, \code{\link{testList}};
  \code{\link{assertLogical}}, \code{\link{checkLogical}},
  \code{\link{testLogical}}; \code{\link{assertMatrix}},
  \code{\link{checkMatrix}}, \code{\link{testMatrix}};
  \code{\link{assertNumeric}}, \code{\link{checkNumeric}},
  \code{\link{testNumeric}}; \code{\link{assertVector}},
  \code{\link{checkVector}}, \code{\link{testVector}}
}

