% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkDirectory}
\alias{assertDirectory}
\alias{checkDirectory}
\alias{testDirectory}
\title{Check for existence and access rights of directories}
\usage{
checkDirectory(x, access = "")

assertDirectory(x, access = "", .var.name)

testDirectory(x, access = "", .var.name)
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}

\item{access}{[\code{character}]\cr
Single string with characters \sQuote{r}, \sQuote{w} and \sQuote{x} to
force a check for read, write or execute access rights.}
}
\value{
Depending on the function prefix:
 If the check is successfull, all functions return \code{TRUE}.
 If the check is not successfull, \code{assertDirectory} throws an error message,
 \code{testDirectory} returns \code{FALSE} and \code{checkDirectory} returns
 a \code{\link[base]{sprintf}} suitable message template with a placeholder
 for the variable name.
}
\description{
Check for existence and access rights of directories
}
\examples{
# Is R's home directory readable?
 test(R.home(), "access", "r")

 # Is R's home directory writable?
 test(R.home(), "access", "w")
}
\seealso{
Other filesystem: \code{\link{assertAccess}},
  \code{\link{checkAccess}}, \code{\link{testAccess}};
  \code{\link{assertFile}}, \code{\link{checkFile}},
  \code{\link{testFile}}
}

