% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIds.R
\name{checkIds}
\alias{checkIds}
\alias{check_ids}
\alias{assertIds}
\alias{assert_ids}
\alias{testIds}
\alias{test_ids}
\alias{expect_ids}
\title{Check strings to be regular identifiers}
\usage{
checkIds(x, unique = FALSE, len = NULL, min.len = NULL)

check_ids(x, unique = FALSE, len = NULL, min.len = NULL)

assertIds(x, unique = FALSE, len = NULL, min.len = NULL,
  .var.name = vname(x), add = NULL)

assert_ids(x, unique = FALSE, len = NULL, min.len = NULL,
  .var.name = vname(x), add = NULL)

testIds(x, unique = FALSE, len = NULL, min.len = NULL)

test_ids(x, unique = FALSE, len = NULL, min.len = NULL)

expect_ids(x, unique = FALSE, len = NULL, min.len = NULL,
  info = NULL, label = vname(x))
}
\arguments{
\item{x}{[\code{character} || \code{NULL}]\cr
Ids to check.}

\item{unique}{[logical(1)]\cr
Must the ids be unique? Default is \code{FALSE}.}

\item{len}{[integer(1)]\cr
Exact length.}

\item{min.len}{[integer(1)]\cr
Minimum length.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in \code{\link{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link{AssertCollection}}.}

\item{info}{[character(1)]\cr
Extra information to be included in the message for the testthat reporter.
See \code{\link[testthat]{expect_that}}.}

\item{label}{[\code{character(1)}]\cr
Name of the checked object to print in messages. Defaults to
the heuristic implemented in \code{\link{vname}}.}
}
\value{
Depending on the function prefix:
 If the check is successful, the functions 
 \code{assertIds}/\code{assert_ids} return 
 \code{x} invisibly, whereas
 \code{checkIds}/\code{check_ids} and 
 \code{testIds}/\code{test_ids} return 
 \code{TRUE}.
 If the check is not successful, 
 \code{assertIds}/\code{assert_ids}
 throws an error message, 
 \code{testIds}/\code{test_ids}
 returns \code{FALSE},
 and \code{checkIds} returns a string with the error message.
 The function \code{expect_ids} always returns an
 \code{\link[testthat]{expectation}}.
}
\description{
Similar to \code{\link{checkNames}}, but more geared towards string identifiers.
Valid Ids are non-empty strings complying to the regular expression pattern
\dQuote{^[.]*[a-zA-Z]+[a-zA-Z0-9._]*$}. This ensures that such ids can safely be
used as variable names, column names or in formulas.
}
\examples{
x = c("foo", "bar")
checkIds(x)

checkIds(c(x, x), unique = TRUE)
}
\seealso{
Other attributes: \code{\link{checkClass}},
  \code{\link{checkMultiClass}}, \code{\link{checkNamed}},
  \code{\link{checkNames}}
}
\concept{attributes}
