% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkInconstant}
\alias{assertInconstant}
\alias{checkInconstant}
\alias{testInconstant}
\title{Check if elements of a vector are inequal / inconstant}
\usage{
checkInconstant(x, tol = .Machine$double.eps^0.5)

assertInconstant(x, tol = .Machine$double.eps^0.5, .var.name)

testInconstant(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}

\item{tol}{[\code{double(1)}]\cr
Numerical tolerance used if \code{x} is of type \code{double} or \code{complex}.
Default is \code{sqrt(.Machine$double.eps)}.}
}
\value{
Depending on the function prefix:
 If the check is successfull, all functions return \code{TRUE}.
 If the check is not successfull, \code{assertInconstant} throws an error message,
 \code{testInconstant} returns \code{FALSE} and \code{checkInconstant} returns
 a string with the error message.
}
\description{
Note that zero-length input is treated as constant.
}
\examples{
test(c(1, NA), "inconstant")
 test(c(0, 1-0.9-0.1), "inconstant")
 test(c(0, 1-0.9-0.1), "inconstant", tol = 0)
}
\seealso{
Other constant: \code{\link{assertConstant}},
  \code{\link{checkConstant}}, \code{\link{testConstant}}
}

