% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkList}
\alias{assertList}
\alias{checkList}
\alias{testList}
\title{Check if an argument is a list}
\usage{
checkList(x, types = character(0L), ...)

assertList(x, types = character(0L), ..., .var.name)

testList(x, types = character(0L), ...)
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}

\item{...}{[ANY]\cr
Additional parameters used in a call of \code{\link{checkVector}}.}

\item{types}{[\code{character}]\cr
Character vector of class names. Each list element must inherit
from at least one of the provided types.
Defaults to \code{character(0)} (no check).}
}
\value{
Depending on the function prefix:
 If the check is successfull, all functions return \code{TRUE}.
 If the check is not successfull, \code{assertList} throws an error message,
 \code{testList} returns \code{FALSE} and \code{checkList} returns
 a string with the error message.
}
\description{
Check if an argument is a list
}
\examples{
test(list(), "list")
 test(as.list(iris), "list", types = c("numeric", "factor"))
}
\seealso{
Other basetypes: \code{\link{assertCharacter}},
  \code{\link{checkCharacter}},
  \code{\link{testCharacter}}; \code{\link{assertComplex}},
  \code{\link{checkComplex}}, \code{\link{testComplex}};
  \code{\link{assertDataFrame}},
  \code{\link{checkDataFrame}},
  \code{\link{testDataFrame}};
  \code{\link{assertEnvironment}},
  \code{\link{checkEnvironment}},
  \code{\link{testEnvironment}};
  \code{\link{assertFactor}}, \code{\link{checkFactor}},
  \code{\link{testFactor}}; \code{\link{assertFunction}},
  \code{\link{checkFunction}}, \code{\link{testFunction}};
  \code{\link{assertIntegerish}},
  \code{\link{checkIntegerish}},
  \code{\link{testIntegerish}};
  \code{\link{assertInteger}}, \code{\link{checkInteger}},
  \code{\link{testInteger}}; \code{\link{assertLogical}},
  \code{\link{checkLogical}}, \code{\link{testLogical}};
  \code{\link{assertMatrix}}, \code{\link{checkMatrix}},
  \code{\link{testMatrix}}; \code{\link{assertNumeric}},
  \code{\link{checkNumeric}}, \code{\link{testNumeric}};
  \code{\link{assertVector}}, \code{\link{checkVector}},
  \code{\link{testVector}}
}

