% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLogical.R
\name{checkLogical}
\alias{checkLogical}
\alias{check_logical}
\alias{assertLogical}
\alias{assert_logical}
\alias{testLogical}
\alias{test_logical}
\alias{expect_logical}
\title{Check if an argument is a vector of type logical}
\usage{
checkLogical(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  null.ok = FALSE)

check_logical(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  null.ok = FALSE)

assertLogical(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  null.ok = FALSE, .var.name = vname(x), add = NULL)

assert_logical(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  null.ok = FALSE, .var.name = vname(x), add = NULL)

testLogical(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  null.ok = FALSE)

test_logical(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  null.ok = FALSE)

expect_logical(x, any.missing = TRUE, all.missing = TRUE, len = NULL,
  min.len = NULL, max.len = NULL, unique = FALSE, names = NULL,
  null.ok = FALSE, info = NULL, label = vname(x))
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{any.missing}{[\code{logical(1)}]\cr
Are vectors with missing values allowed? Default is \code{TRUE}.}

\item{all.missing}{[\code{logical(1)}]\cr
Are vectors with only missing values allowed? Default is \code{TRUE}.}

\item{len}{[\code{integer(1)}]\cr
Exact expected length of \code{x}.}

\item{min.len}{[\code{integer(1)}]\cr
Minimal length of \code{x}.}

\item{max.len}{[\code{integer(1)}]\cr
Maximal length of \code{x}.}

\item{unique}{[\code{logical(1)}]\cr
Must all values be unique? Default is \code{FALSE}.}

\item{names}{[\code{character(1)}]\cr
Check for names. See \code{\link{checkNamed}} for possible values.
Default is \dQuote{any} which performs no check at all.
Note that you can use \code{\link{checkSubset}} to check for a specific set of names.}

\item{null.ok}{[\code{logical(1)}]\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in \code{\link{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link{AssertCollection}}.}

\item{info}{[character(1)]\cr
Extra information to be included in the message for the testthat reporter.
See \code{\link[testthat]{expect_that}}.}

\item{label}{[\code{character(1)}]\cr
Name of the checked object to print in messages. Defaults to
the heuristic implemented in \code{\link{vname}}.}
}
\value{
Depending on the function prefix:
 If the check is successful, the functions 
 \code{assertLogical}/\code{assert_logical} return 
 \code{x} invisibly, whereas
 \code{checkLogical}/\code{check_logical} and 
 \code{testLogical}/\code{test_logical} return 
 \code{TRUE}.
 If the check is not successful, 
 \code{assertLogical}/\code{assert_logical}
 throws an error message, 
 \code{testLogical}/\code{test_logical}
 returns \code{FALSE},
 and \code{checkLogical} returns a string with the error message.
 The function \code{expect_logical} always returns an
 \code{\link[testthat]{expectation}}.
}
\description{
Check if an argument is a vector of type logical
}
\details{
This function does not distinguish between
\code{NA}, \code{NA_integer_}, \code{NA_real_}, \code{NA_complex_}
\code{NA_character_} and \code{NaN}.
}
\examples{
testLogical(TRUE)
testLogical(TRUE, min.len = 1)
}
\seealso{
\code{\link{checkBit}}

Other basetypes: \code{\link{checkArray}},
  \code{\link{checkAtomicVector}},
  \code{\link{checkAtomic}}, \code{\link{checkCharacter}},
  \code{\link{checkComplex}}, \code{\link{checkDataFrame}},
  \code{\link{checkDate}}, \code{\link{checkDouble}},
  \code{\link{checkEnvironment}},
  \code{\link{checkFactor}}, \code{\link{checkFormula}},
  \code{\link{checkFunction}},
  \code{\link{checkIntegerish}},
  \code{\link{checkInteger}}, \code{\link{checkList}},
  \code{\link{checkMatrix}}, \code{\link{checkNull}},
  \code{\link{checkNumeric}}, \code{\link{checkPOSIXct}},
  \code{\link{checkRaw}}, \code{\link{checkVector}}
}
\concept{basetypes}
