% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{checkmate-package}
\alias{checkmate}
\alias{checkmate-package}
\title{Fast and Versatile Argument Checks}
\description{
Tests and assertions to perform frequent argument checks. A
substantial part of the package was written in C to minimize any worries
about execution time overhead.
}
\details{
\describe{
  \item{Homepage:}{\url{https://github.com/mllg/checkmate}}
}
}

