% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{checkmate}
\alias{checkmate}
\alias{checkmate-package}
\title{The checkmate package}
\description{
The checkmate package
}
\section{Additional information}{


\describe{
  \item{Homepage:}{\url{https://github.com/mllg/checkmate}}
}
}

