% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkNull}
\alias{checkNull}
\alias{testNull}
\title{Check if an argument is NULL}
\usage{
checkNull(x)

testNull(x)
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}
}
\value{
Depending on the function prefix:
 If the check is successfull, all functions return \code{TRUE}.
 If the check is not successfull, \code{assertNull} throws an error message,
 \code{testNull} returns \code{FALSE} and \code{checkNull} returns
 a string with the error message.
}
\description{
Check if an argument is NULL
}

