% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkNumber}
\alias{checkNumber}
\title{Check if an argument is single numeric}
\usage{
checkNumber(x, na.ok = FALSE)
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}

\item{na.ok}{[\code{logical(1)}]\cr
Are missing values allowed? Default is \code{FALSE}.}
}
\value{
Depending on the function prefix:
 If the check is successfull, all functions return \code{TRUE}.
 If the check is not successfull, \code{assertCount} throws an error message,
 \code{testCount} returns \code{FALSE} and \code{checkCount} returns
 a string with the error message.
}
\description{
Check if an argument is single numeric
}
\details{
This function does not distinguish between
\code{NA}, \code{NA_integer_}, \code{NA_real_}, \code{NA_complex_}
and \code{NA_character_}.
}
\examples{
test(1, "count")
 test(-1, "count")
 test(Inf, "count")
}

