% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogenerated.r, R/checkPercentage.r
\name{assertPercentage}
\alias{assertPercentage}
\alias{assert_percentage}
\alias{checkPercentage}
\alias{expect_percentage}
\alias{testPercentage}
\alias{test_percentage}
\title{Check if an argument is a percentage}
\usage{
assertPercentage(x, na.ok = FALSE, add = NULL, .var.name)

assert_percentage(x, na.ok = FALSE, add = NULL, .var.name)

testPercentage(x, na.ok = FALSE)

test_percentage(x, na.ok = FALSE)

expect_percentage(x, na.ok = FALSE, info = NULL, label = NULL)

checkPercentage(x, na.ok = FALSE)
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{na.ok}{[\code{logical(1)}]\cr
Are missing values allowed? Default is \code{FALSE}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertions. See \code{\link{AssertCollection}}.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}

\item{info}{[character(1)]\cr
Extra information to be included in the message for the testthat reporter.
See \code{\link[testthat]{expect_that}}.}

\item{label}{[character(1)]\cr
Same as \code{.var.name}, but passed down to \code{\link[testthat]{expect_that}}.}
}
\value{
Depending on the function prefix:
 If the check is successful, all functions return \code{TRUE}. If the check
 is not successful, \code{assertPercentage}/\code{assert_percentage}
 throws an error message, \code{testPercentage}/\code{test_percentage}
 returns \code{FALSE},
 and \code{checkPercentage} returns a string with the error message.
 The function \code{expect_percentage} will return an
 \code{\link[testthat]{expectation}}.
}
\description{
This checks \code{x} to be numeric and in the range \code{[0,1]}.
}
\details{
This function does not distinguish between
\code{NA}, \code{NA_integer_}, \code{NA_real_}, \code{NA_complex_}
\code{NA_character_} and \code{NaN}.
}
\examples{
testPercentage(0.5)
testPercentage(1)
}
\seealso{
Other scalars: \code{\link{assertCount}},
  \code{\link{assertFlag}}, \code{\link{assertInt}},
  \code{\link{assertNumber}}, \code{\link{assertScalarNA}},
  \code{\link{assertScalar}}, \code{\link{assertString}}
}

