% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkSubset}
\alias{assertSubset}
\alias{checkSubset}
\alias{testSubset}
\title{Check if object is a subset of a given set}
\usage{
checkSubset(x, choices)

assertSubset(x, choices, .var.name)

testSubset(x, choices)
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}

\item{choices}{[\code{atomic}]\cr
Set of possible values.}
}
\value{
Depending on the function prefix:
 If the check is successfull, all functions return \code{TRUE}.
 If the check is not successfull, \code{assertSubset} throws an error message,
 \code{testSubset} returns \code{FALSE} and \code{checkSubset} returns
 a \code{\link[base]{sprintf}} suitable message template with a placeholder
 for the variable name.
}
\description{
Check if object is a subset of a given set
}
\examples{
test(c("a", "z"), "subset", letters)
}
\seealso{
Other set: \code{\link{assertChoice}},
  \code{\link{checkChoice}}, \code{\link{testChoice}}
}

