% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTibble.R
\name{checkTibble}
\alias{assertTibble}
\alias{assert_tibble}
\alias{checkTibble}
\alias{check_tibble}
\alias{expect_tibble}
\alias{testTibble}
\alias{test_tibble}
\title{Check if an argument is a tibble}
\usage{
checkTibble(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE)

check_tibble(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE)

assertTibble(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE,
  .var.name = vname(x), add = NULL)

assert_tibble(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE,
  .var.name = vname(x), add = NULL)

testTibble(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE)

test_tibble(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE)

expect_tibble(x, types = character(0L), any.missing = TRUE,
  all.missing = TRUE, min.rows = NULL, min.cols = NULL, nrows = NULL,
  ncols = NULL, row.names = NULL, col.names = NULL, null.ok = FALSE,
  info = NULL, label = vname(x))
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{types}{[\code{character}]\cr
Character vector of class names. Each list element must inherit
from at least one of the provided types.
The types \dQuote{logical}, \dQuote{integer}, \dQuote{integerish}, \dQuote{double},
\dQuote{numeric}, \dQuote{complex}, \dQuote{character}, \dQuote{factor}, \dQuote{atomic}, \dQuote{vector}
\dQuote{atomicvector}, \dQuote{array}, \dQuote{matrix}, \dQuote{list}, \dQuote{function},
\dQuote{environment} and \dQuote{null} are supported.
For other types \code{\link[base]{inherits}} is used as a fallback to check \code{x}'s inheritance.
Defaults to \code{character(0)} (no check).}

\item{any.missing}{[\code{logical(1)}]\cr
Are missing values allowed? Default is \code{TRUE}.}

\item{all.missing}{[\code{logical(1)}]\cr
Are matrices with only missing values allowed? Default is \code{TRUE}.}

\item{min.rows}{[\code{integer(1)}]\cr
Minimum number of rows.}

\item{min.cols}{[\code{integer(1)}]\cr
Minimum number of columns.}

\item{nrows}{[\code{integer(1)}]\cr
Exact number of rows.}

\item{ncols}{[\code{integer(1)}]\cr
Exact number of columns.}

\item{row.names}{[\code{character(1)}]\cr
Check for row names. Default is \dQuote{NULL} (no check).
See \code{\link{checkNamed}} for possible values.
Note that you can use \code{\link{checkSubset}} to check for a specific set of names.}

\item{col.names}{[\code{character(1)}]\cr
Check for column names. Default is \dQuote{NULL} (no check).
See \code{\link{checkNamed}} for possible values.
Note that you can use \code{\link{checkSubset}} to test for a specific set of names.}

\item{null.ok}{[\code{logical(1)}]\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in \code{\link{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link{AssertCollection}}.}

\item{info}{[character(1)]\cr
Extra information to be included in the message for the testthat reporter.
See \code{\link[testthat]{expect_that}}.}

\item{label}{[\code{character(1)}]\cr
Name of the checked object to print in messages. Defaults to
the heuristic implemented in \code{\link{vname}}.}
}
\value{
Depending on the function prefix:
 If the check is successful, the functions 
 \code{assertTibble}/\code{assert_tibble} return 
 \code{x} invisibly, whereas
 \code{checkTibble}/\code{check_tibble} and 
 \code{testTibble}/\code{test_tibble} return 
 \code{TRUE}.
 If the check is not successful, 
 \code{assertTibble}/\code{assert_tibble}
 throws an error message, 
 \code{testTibble}/\code{test_tibble}
 returns \code{FALSE},
 and \code{checkTibble} returns a string with the error message.
 The function \code{expect_tibble} always returns an
 \code{\link[testthat]{expectation}}.
}
\description{
Check if an argument is a tibble
}
\examples{
library(tibble)
x = as_tibble(iris)
testTibble(x)
testTibble(x, nrow = 150, any.missing = FALSE)
}
\seealso{
Other compound: \code{\link{checkArray}},
  \code{\link{checkDataFrame}},
  \code{\link{checkDataTable}}, \code{\link{checkMatrix}}
}

