% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coalesce.r
\name{\%??\%}
\alias{\%??\%}
\title{Coalesce operator}
\usage{
lhs \%??\% rhs
}
\description{
Returns the left hand side if not missing nor \code{NULL}, and
the right hand side otherwise.
}
\examples{
print(NULL \%??\% 1 \%??\% 2)
print(names(iris) \%??\% letters[seq_len(ncol(iris))])
}

