% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectAssertion.R
\name{inspectAssertion}
\alias{inspectAssertion}
\title{Inspect an Assertion}
\usage{
inspectAssertion(expr)
}
\arguments{
\item{expr}{[\code{expression}]\cr
Code to execute.}
}
\value{
[\code{list}] with named elements \dQuote{message} and \dQuote{x} if an assertion
 has been caught, and the return value of \code{expr} otherwise.
}
\description{
Executes the expression \code{expr} and uses \code{\link[base]{tryCatch}} to catch
the first \code{\link[conditions]{assertion_error}} raised.
}
\examples{
getElement = function(x, i) {
  assertAtomicVector(x, any.missing = FALSE)
  assertCount(i)
  x[i]
}
inspectAssertion(getElement(1:10, 2))
inspectAssertion(getElement(NULL, 1))
inspectAssertion(getElement(iris[1:3, ], 1))
inspectAssertion(getElement(1:10, -2))
}
