% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.R
\name{.setequal}
\alias{.equal}
\alias{.in}
\alias{.setequal}
\title{Sets of Names}
\usage{
.setequal(x)

.in(x)

.equal(x)
}
\arguments{
\item{x}{[\code{character}]\cr
Character vector of names.}
}
\value{
[\code{character}]. Returns \code{x} with an attribute attached.
}
\description{
Little helper function to be used if you want to check that the names of an object are setequal,
equal or a subset of certain names.
}
\examples{
x = setNames(1:3, letters[1:3])
checkNumeric(x, names = .in(letters))
checkNumeric(x, names = .setequal(letters[1:3]))
checkDataFrame(iris, col.names = .setequal(sample(names(iris))))
}

