% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExpectation.R
\name{register_test_backend}
\alias{register_test_backend}
\title{Select Backend for Unit Tests}
\usage{
register_test_backend(name)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
\code{"testthat"} or \code{"tinytest"}.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Allows to explicitly select a backend for the unit tests.
Currently supported are \code{"testthat"} and \code{"tinytest"}.
The respective package must be installed and are loaded (but not attached).

If this function is not explicitly called, defaults to \code{"testthat"} unless
the \code{"tinytest"}'s namespace is loaded.
}
