% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vname.R
\name{vname}
\alias{vname}
\title{Lookup a variable name}
\usage{
vname(x)
}
\arguments{
\item{x}{[ANY]\cr
Object.}
}
\value{
[\code{character(1)}] Variable name.
}
\description{
Tries to heuristically determine the variable name of \code{x} in the parent frame
with a combination of \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.
Used for checkmate's error messages.
}
