% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anyNaN.r
\name{anyNaN}
\alias{anyNaN}
\title{Check if an object contains NaN values}
\usage{
anyNaN(x)
}
\arguments{
\item{x}{[\code{ANY}]\cr
Object to check.}
}
\value{
[\code{logical(1)}] Returns \code{TRUE} if any element if \code{NaN}.
}
\description{
Supported are atomic types (see \code{\link[base]{is.atomic}}), lists and data frames.
}
\examples{
anyNaN(1:10)
anyNaN(c(1:10, NaN))
iris[3, 3] = NaN
anyNaN(iris)
}

