% Generated by roxygen2 (4.0.1): do not edit by hand
\name{assert}
\alias{assert}
\title{Combine multiple checks to an assertion}
\usage{
assert(..., .var.name)
}
\arguments{
\item{...}{[ANY]\cr
List of calls to check functions.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name.}
}
\value{
Throws an error if all checks failed and invisibly returns
 \code{TRUE} otherwise.
}
\description{
Combine multiple checks to an assertion
}

