% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_data.frame}
\alias{check_data.frame}
\title{Check if an argument is a \code{data.frame}}
\usage{
check_data.frame(x, types = character(0L), any.missing = TRUE, min.rows,
  min.cols, nrows, ncols, row.names = "any", col.names = "any")
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{any.missing}{[\code{integer(1)}]\cr
Are missing values allowed? Default is \code{TRUE}.}

\item{min.rows}{[\code{integer(1)}]\cr
Minimum number of rows.}

\item{min.cols}{[\code{integer(1)}]\cr
Minimum number of columns.}

\item{nrows}{[\code{integer(1)}]\cr
Exact number of rows.}

\item{ncols}{[\code{integer(1)}]\cr
Exact number of columns.}

\item{row.names}{[\code{logical(1)}]\cr
Check for row names. Default is \dQuote{any} (no check).
See \code{\link{check_named}} for possible values.}

\item{col.names}{[\code{logical(1)}]\cr
Check for column names. Default is \dQuote{any} (no check).
See \code{\link{check_named}} for possible values.}

\item{types}{[\code{character}]\cr
Character vector of class names. Each list element must inherit
from at least one of the provided types.
Defaults to \code{character(0)} (no check).}
}
\value{
Returns \code{TRUE} on success and an error message otherwise.
 Intended to be used with \code{\link{test}}, \code{\link{assert}} or \code{\link{pass}}.
}
\description{
Check if an argument is a \code{data.frame}
}
\examples{
test(iris, "data.frame")
 test(iris, "data.frame", min.rows = 1, col.names = "named")
}
\seealso{
Other basetypes: \code{\link{check_character}};
  \code{\link{check_complex}};
  \code{\link{check_environment}};
  \code{\link{check_factor}}; \code{\link{check_function}};
  \code{\link{check_integerish}};
  \code{\link{check_integer}}; \code{\link{check_list}};
  \code{\link{check_logical}}; \code{\link{check_matrix}};
  \code{\link{check_numeric}}; \code{\link{check_vector}}
}

