% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_named}
\alias{check_named}
\title{Check if an argument is named}
\usage{
check_named(x, type = "named")
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{type}{[character(1)]\cr
Select the check(s) to perform.
\dQuote{unnamed} checks \code{x} to be unnamed.
\dQuote{named} (default) checks \code{x} to be named, this includes names to be not \code{NA} or emtpy (\code{""}).
\dQuote{unique} additionally tests for non-duplicated names and \code{strict} tests the names to comply to R's variable naming rules on top of \dQuote{unique}.
Note that for zero-length \code{x} every name check evalutes to \code{TRUE}.}
}
\value{
Returns \code{TRUE} on success and an error message otherwise.
 Intended to be used with \code{\link{test}}, \code{\link{assert}} or \code{\link{pass}}.
}
\description{
Check if an argument is named
}
\examples{
x = 1:3
 test(x, "named", "unnamed")
 names(x) = letters[1:3]
 test(x, "named", "unique")
}

