% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkArray.R
\name{checkArray}
\alias{checkArray}
\alias{check_array}
\alias{assertArray}
\alias{assert_array}
\alias{testArray}
\alias{test_array}
\alias{expect_array}
\title{Check if an argument is an array}
\usage{
checkArray(x, mode = NULL, any.missing = TRUE, d = NULL, min.d = NULL,
  max.d = NULL, null.ok = FALSE)

check_array(x, mode = NULL, any.missing = TRUE, d = NULL, min.d = NULL,
  max.d = NULL, null.ok = FALSE)

assertArray(x, mode = NULL, any.missing = TRUE, d = NULL, min.d = NULL,
  max.d = NULL, null.ok = FALSE, .var.name = vname(x), add = NULL)

assert_array(x, mode = NULL, any.missing = TRUE, d = NULL, min.d = NULL,
  max.d = NULL, null.ok = FALSE, .var.name = vname(x), add = NULL)

testArray(x, mode = NULL, any.missing = TRUE, d = NULL, min.d = NULL,
  max.d = NULL, null.ok = FALSE)

test_array(x, mode = NULL, any.missing = TRUE, d = NULL, min.d = NULL,
  max.d = NULL, null.ok = FALSE)

expect_array(x, mode = NULL, any.missing = TRUE, d = NULL, min.d = NULL,
  max.d = NULL, null.ok = FALSE, info = NULL, label = vname(x))
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{mode}{[\code{character(1)}]\cr
Storage mode of the array. Arrays can hold vectors, i.e. \dQuote{logical},
\dQuote{integer}, \dQuote{integerish}, \dQuote{double}, \dQuote{numeric}, \dQuote{complex},
\dQuote{character} and \dQuote{list}. You can also specify \dQuote{atomic}
here to explicitly prohibit lists. Default is \code{NULL} (no check).}

\item{any.missing}{[\code{logical(1)}]\cr
Are missing values allowed? Default is \code{TRUE}.}

\item{d}{[\code{integer(1)}]\cr
Exact number of dimensions of array \code{x}.
Default is \code{NULL} (no check).}

\item{min.d}{[\code{integer(1)}]\cr
Minimum number of dimensions of array \code{x}.
Default is \code{NULL} (no check).}

\item{max.d}{[\code{integer(1)}]\cr
Maximum number of dimensions of array \code{x}.
Default is \code{NULL} (no check).}

\item{null.ok}{[\code{logical(1)}]\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in \code{\link{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link{AssertCollection}}.}

\item{info}{[character(1)]\cr
Extra information to be included in the message for the testthat reporter.
See \code{\link[testthat]{expect_that}}.}

\item{label}{[\code{character(1)}]\cr
Name of the checked object to print in messages. Defaults to
the heuristic implemented in \code{\link{vname}}.}
}
\value{
Depending on the function prefix:
 If the check is successful, the functions 
 \code{assertArray}/\code{assert_array} return 
 \code{x} invisibly, whereas
 \code{checkArray}/\code{check_array} and 
 \code{testArray}/\code{test_array} return 
 \code{TRUE}.
 If the check is not successful, 
 \code{assertArray}/\code{assert_array}
 throws an error message, 
 \code{testArray}/\code{test_array}
 returns \code{FALSE},
 and \code{checkArray} returns a string with the error message.
 The function \code{expect_array} always returns an
 \code{\link[testthat]{expectation}}.
}
\description{
Check if an argument is an array
}
\examples{
checkArray(array(1:27, dim = c(3, 3, 3)), d = 3)
}
\seealso{
Other basetypes: \code{\link{checkAtomicVector}},
  \code{\link{checkAtomic}}, \code{\link{checkCharacter}},
  \code{\link{checkComplex}}, \code{\link{checkDataFrame}},
  \code{\link{checkDate}}, \code{\link{checkEnvironment}},
  \code{\link{checkFactor}}, \code{\link{checkFunction}},
  \code{\link{checkIntegerish}},
  \code{\link{checkInteger}}, \code{\link{checkList}},
  \code{\link{checkLogical}}, \code{\link{checkMatrix}},
  \code{\link{checkNull}}, \code{\link{checkNumeric}},
  \code{\link{checkVector}}

Other compound: \code{\link{checkDataFrame}},
  \code{\link{checkDataTable}}, \code{\link{checkMatrix}},
  \code{\link{checkTibble}}
}
