% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkChoice.r
\name{checkChoice}
\alias{assertChoice}
\alias{assert_choice}
\alias{checkChoice}
\alias{expect_choice}
\alias{testChoice}
\alias{test_choice}
\title{Check if an object is an element of a given set}
\usage{
checkChoice(x, choices)

assertChoice(x, choices, add = NULL, .var.name)

assert_choice(x, choices, add = NULL, .var.name)

testChoice(x, choices)

test_choice(x, choices)

expect_choice(x, choices, info = NULL, label = NULL)
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{choices}{[\code{atomic}]\cr
Set of possible values.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertions. See \code{\link{AssertCollection}}.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}

\item{info}{[character(1)]\cr
Extra information to be included in the message for the testthat reporter.
See \code{\link[testthat]{expect_that}}.}

\item{label}{[character(1)]\cr
Same as \code{.var.name}, but passed down to \code{\link[testthat]{expect_that}}.}
}
\value{
Depending on the function prefix:
 If the check is successful, the functions return \code{TRUE}. If the check
 is not successful, \code{assertChoice}/\code{assert_choice}
 throws an error message, \code{testChoice}/\code{test_choice}
 returns \code{FALSE},
 and \code{checkChoice} returns a string with the error message.
 The function \code{expect_choice} always returns an
 \code{\link[testthat]{expectation}}.
}
\description{
Check if an object is an element of a given set
}
\examples{
testChoice("x", letters)
}
\seealso{
Other set: \code{\link{checkSetEqual}},
  \code{\link{checkSubset}}
}

