% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkClass.R
\name{checkClass}
\alias{assertClass}
\alias{assert_class}
\alias{checkClass}
\alias{expect_class}
\alias{testClass}
\alias{test_class}
\title{Check the class membership of an argument}
\usage{
checkClass(x, classes, ordered = FALSE)

assertClass(x, classes, ordered = FALSE, .var.name = vname(x), add = NULL)

assert_class(x, classes, ordered = FALSE, .var.name = vname(x),
  add = NULL)

testClass(x, classes, ordered = FALSE)

test_class(x, classes, ordered = FALSE)

expect_class(x, classes, ordered = FALSE, info = NULL, label = vname(x))
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{classes}{[\code{character}]\cr
Class names to check for inheritance with \code{\link[base]{inherits}}.}

\item{ordered}{[\code{logical(1)}]\cr
Expect \code{x} to be specialized in provided order.
Default is \code{FALSE}.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in \code{\link{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link{AssertCollection}}.}

\item{info}{[character(1)]\cr
Extra information to be included in the message for the testthat reporter.
See \code{\link[testthat]{expect_that}}.}

\item{label}{[\code{character(1)}]\cr
Name of the checked object to print in messages. Defaults to
the heuristic implemented in \code{\link{vname}}.}
}
\value{
Depending on the function prefix:
 If the check is successful, the functions return \code{TRUE}. If the check
 is not successful, \code{assertClass}/\code{assert_class}
 throws an error message, \code{testClass}/\code{test_class}
 returns \code{FALSE},
 and \code{checkClass} returns a string with the error message.
 The function \code{expect_class} always returns an
 \code{\link[testthat]{expectation}}.
}
\description{
Check the class membership of an argument
}
\examples{
# Create an object with classes "foo" and "bar"
x = 1
class(x) = c("foo", "bar")

# is x of class "foo"?
testClass(x, "foo")

# is x of class "foo" and "bar"?
testClass(x, c("foo", "bar"))

# is x most specialized as "bar"?
testClass(x, "bar", ordered = TRUE)
}

