% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkCount}
\alias{assertCount}
\alias{checkCount}
\alias{testCount}
\title{Check if an argument is a count}
\usage{
checkCount(x, na.ok = FALSE, positive = FALSE)

assertCount(x, na.ok = FALSE, positive = FALSE, .var.name)

testCount(x, na.ok = FALSE, positive = FALSE)
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}

\item{na.ok}{[\code{logical(1)}]\cr
Are missing values allowed? Default is \code{FALSE}.}

\item{positive}{[\code{logical(1)}]\cr
Must \code{x} be positive (>= 1)?
Default is \code{FALSE}.}
}
\value{
Depending on the function prefix:
 If the check is successfull, all functions return \code{TRUE}.
 If the check is not successfull, \code{assertCount} throws an error message,
 \code{testCount} returns \code{FALSE} and \code{checkCount} returns
 a string with the error message.
}
\description{
A count is a non-negative integer.
}
\details{
This function does not distinguish between
\code{NA}, \code{NA_integer_}, \code{NA_real_}, \code{NA_complex_}
and \code{NA_character_}.
}
\examples{
testCount(1)
 testCount(-1)
}

