% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkFile}
\alias{assertFile}
\alias{checkFile}
\alias{testFile}
\title{Check existence and access rights of files}
\usage{
checkFile(x, access = "")

assertFile(x, access = "", .var.name)

testFile(x, access = "")
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}

\item{access}{[\code{character}]\cr
Single string with characters \sQuote{r}, \sQuote{w} and \sQuote{x} to
force a check for read, write or execute access rights.}
}
\value{
Depending on the function prefix:
 If the check is successfull, all functions return \code{TRUE}.
 If the check is not successfull, \code{assertFile} throws an error message,
 \code{testFile} returns \code{FALSE} and \code{checkFile} returns
 a \code{\link[base]{sprintf}} suitable message template with a placeholder
 for the variable name.
}
\description{
Check existence and access rights of files
}
\examples{
# Check if R's COPYING file is readable
 test(file.path(R.home(), "COPYING"), "file", access = "r")

 # Check if R's COPYING file is writable
 test(file.path(R.home(), "COPYING"), "file", access = "w")
}
\seealso{
Other filesystem: \code{\link{assertAccess}},
  \code{\link{checkAccess}}, \code{\link{testAccess}};
  \code{\link{assertDirectory}},
  \code{\link{checkDirectory}}, \code{\link{testDirectory}}
}

