% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{checkmate}
\alias{checkmate}
\alias{checkmate-package}
\title{The checkmate package}
\description{
The checkmate package
}
\section{Additional information}{


\describe{
  \item{Homepage:}{\url{https://github.com/mllg/checkmate}}
}
}

