\name{checkNamed}
\alias{assertNamed}
\alias{checkNamed}
\title{Check or assert that an argument is named}
\usage{
checkNamed(x)

assertNamed(x)
}
\arguments{
  \item{x}{[ANY]\cr Object to check.}
}
\value{
[\code{logical(1)}] Returns \code{TRUE} on success.  Throws
an exception on failure for assertion.
}
\description{
Check or assert that an argument is named
}

