% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkNames}
\alias{assertNames}
\alias{checkNames}
\alias{testNames}
\title{Check for names}
\usage{
checkNames(x, type = "named")

assertNames(x, type = "named", .var.name)

testNames(x, type = "named")
}
\arguments{
\item{x}{[ANY]\cr
Object to check.}

\item{type}{[character(1)]\cr
Select the check(s) to perform.
\dQuote{unnamed} checks \code{x} to be \code{NULL}.
\dQuote{named} (default) checks \code{x} for reguluar names which excludes names to be \code{NA} or emtpy (\code{""}).
\dQuote{unique} additionally tests for non-duplicated names.
\dQuote{strict} checks for unique names which comply to R's variable name restrictions.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}
}
\value{
Depending on the function prefix:
 If the check is successful, all functions return \code{TRUE}.
 If the check is not successful, \code{assertNamed} throws an error message,
 \code{testNamed} returns \code{FALSE} and \code{checkNamed} returns
 a string with the error message.
}
\description{
Check for names
}
\examples{
x = 1:3
testNames(x, "unnamed")
names(x) = letters[1:3]
testNames(x, "unique")
}

