% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkNull.r
\name{checkNull}
\alias{assertNull}
\alias{checkNull}
\alias{testNull}
\title{Check if an argument is NULL}
\usage{
checkNull(x)

assertNull(x, .var.name)

testNull(x)
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{.var.name}{[character(1)]\cr
Name for \code{x}. Defaults to a heuristic to determine
the name using \code{\link[base]{deparse}} and \code{\link[base]{substitute}}.}
}
\value{
Depending on the function prefix:
 If the check is successful, all functions return \code{TRUE}.
 If the check is not successful, \code{assertNull} throws an error message,
 \code{testNull} returns \code{FALSE} and \code{checkNull} returns
 a string with the error message.
}
\description{
Check if an argument is NULL
}
\examples{
testNull(NULL)
testNull(1)
}

