% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coalesce.r
\name{\%??\%}
\alias{\%??\%}
\title{Coalesce operator}
\usage{
lhs \%??\% rhs
}
\arguments{
\item{lhs}{[any]\cr
Left hand side of the operator. Is returned, if not missing or \code{NULL}.}

\item{rhs}{[any]\cr
Right hand side of the operator. Is returned, if \code{lhs} is missing or \code{NULL}.}
}
\value{
Either \code{lhs} or \code{rhs}.
}
\description{
Returns the left hand side if not missing nor \code{NULL}, and
the right hand side otherwise.
}
\examples{
print(NULL \%??\% 1 \%??\% 2)
print(names(iris) \%??\% letters[seq_len(ncol(iris))])
}

